+++ CHANGELOG.md
@@ -5,6 +5,7 @@ changes with their rationale when appropriate:
 
 ### v3.243.0 (uncut)
 - [all] Upgrade some dependency versions, including Kotlin to `1.3.72`.
+- [http4k-security-oauth] A strategy can now be passed into `AuthRequestWithRequestAuthRequestExtractor` to determine how to combine `AuthRequest` and `RequestObject` H/T @tom
 
 ### v3.242.0
 - [all] Upgrade some dependency versions.